requirements.txt
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-surprise==1.1.3
scipy==1.11.3
six==1.16.0
tzdata==2023.3
